+++ ChanTicker.Core/Interfaces/ITick.cs
@@ -5,6 +5,6 @@ namespace ChanTicker.Core.Interfaces
     public interface ITick
     {
         DateTimeOffset TimeStamp { get; }
-        double? Price { get; }
+        decimal? Price { get; }
     }
 }
